Security Advisories: Add information from Riscure mini-audit
Out of all potentially security issues reported by Riscure as an outcome
of their "mini-audit" we narrowed down seven of them as higher severity.
These are the ones that has been mentioned in this patch.

CVE numbers have been requested, but we haven't got them yet. If/when we
get those, we will submit another update patch.

Signed-off-by: Joakim Bech <joakim.bech@linaro.org>